handle users w/o account for comments
